Change fee_rate_cmp to accept i64, matching stellar-core parity
Change fee_rate_cmp signature from (u64, u32, u64, u32) to (i64, u32, i64, u32)
to match stellar-core's feeRate3WayCompare(int64_t, ...). Add hard assert for
non-negative fees, matching stellar-core's releaseAssertOrThrow in bigMultiply.

This eliminates two unsound `as u64` casts in parallel_tx_set_builder.rs and
tx_queue_limiter.rs that would silently wrap negative i64 fees to large u64
values, corrupting fee-rate sort order.

Add unit tests for fee_rate_cmp (positive comparison, zero fees, overflow,
panic on negative) and a caller-level regression test proving negative
inclusion fees now panic instead of silently reordering.

Closes #1823

+++ crates/herder/src/surge_pricing.rs
@@ -287,10 +287,11 @@ impl PartialOrd for QueueEntry {
 
 impl Ord for QueueEntry {
     fn cmp(&self, other: &Self) -> Ordering {
+        // inclusion_fee is u64 validated non-negative at construction (from XDR i64).
         let ord = fee_rate_cmp(
-            self.inclusion_fee,
+            self.inclusion_fee as i64,
             self.op_count,
-            other.inclusion_fee,
+            other.inclusion_fee as i64,
             other.op_count,
         );
         if ord != Ordering::Equal {
@@ -704,9 +705,9 @@ impl SurgePricingPriorityQueue {
             };
 
             if fee_rate_cmp(
-                entry.inclusion_fee,
+                entry.inclusion_fee as i64,
                 entry.op_count,
-                tx.inclusion_fee,
+                tx.inclusion_fee as i64,
                 tx.op_count,
             ) != Ordering::Less
             {
